+++ pkg/database/stream.go
@@ -1,0 +1,8 @@
+package database
+
+// Stream is an interface for streaming events from a collection.
+type Stream interface {
+	Next() <-chan Event
+	Done() <-chan struct{}
+	Close() error
+}
